Title: 바이러스 검사, Time: 2015ms, Memory: 159MB, Status: Passed - CodeTree

+++ 240219/바이러스 검사/virus-detector.java
@@ -11,7 +11,7 @@ public class Main {
         int sup = sc.nextInt();
         int sub = sc.nextInt();
 
-        int cnt = N;
+        long cnt = N;
         for(int i = 0 ; i < N ; i++){
             int temp = arr[i];
             temp -= sup;
